+++ src/reportlab/lib/rparsexml.py
@@ -80,12 +80,12 @@ except ImportError:
 class smartDecode:
     @staticmethod
     def __call__(s):
-        print('initial')
+        #print('initial')
         import chardet
         def __call__(s):
             if isinstance(s,str): return s
             cdd = chardet.detect(s)
-            print('final')
+            #print('final')
             return s.decode(cdd["encoding"])
         smartDecode.__class__.__call__ = staticmethod(__call__)
         return  __call__(s)
@@ -413,7 +413,6 @@ def testparse(s,dump=0):
     from time import time
     from pprint import pprint
     now = time()
-    breakpoint()
     D = parsexmlSimple(s,oneOutermostTag=1)
     print("DONE", time()-now)
     if dump&4: