+++ CHANGES.md
@@ -11,6 +11,10 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
+CHANGES  4.0.8   ??/??/2023
+---------------------------
+    * fix rparsexml bug contributed by Matt Folwell  mjf at pearson co uk 
+
 CHANGES  4.0.7   06/11/2023
 ---------------------------
     * allow testutils.makeSuiteForClasses to specify a prefix for test method names